requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
python-dotenv>=1.0.1
google-generativeai>=0.3.0
gspread>=6.0.0
google-auth>=2.29.0